Fixed Intrusive Thought After Game Finish
Before the intrusive thoughts from the previous failed game would carry over now they don't. In addition I removed the system that made sure you couldn't have two spawnpoints in a row that are the same because the system didn't work and the speed increase makes it no longer a problem. Also fixed naming on intrusiveThoughtList and Communsim missspelling.